requirements.txt
python-telegram-bot==20.6
aiohttp
xmltodict
requests
beautifulsoup4